https://review.openocd.org/c/openocd/+/8384/comment/b035f2ec_ce46a82e/

+++ src/flash/nor/mspm0.c
@@ -519,19 +519,19 @@ static int mspm0_fctl_get_sector_reg(struct flash_bank *bank, unsigned int addr,
 
 	if (sector_num < mspm0_info->main_flash_size_kb) {
 		/* Use CMDWEPROTA */
-		if (phys_sector_num < (unsigned int)32) {
-			*sector_mask = (unsigned int)1 << phys_sector_num;
+		if (phys_sector_num < 32) {
+			*sector_mask = 1 << phys_sector_num;
 			*reg = FCTL_REG_CMDWEPROTA;
 			return ERROR_OK;
 		}
 
 		/* Use CMDWEPROTB */
-		if (sector_in_bank < (unsigned int)256) {
+		if (sector_in_bank < 256) {
 			/* Dual bank system */
 			if (mspm0_info->main_flash_num_banks > 1) {
-				*sector_mask = (unsigned int)1 << (sector_in_bank / 8);
+				*sector_mask = 1 << (sector_in_bank / 8);
 			} else {	/* Single bank system */
-				*sector_mask = (unsigned int)1 << ((sector_in_bank - 32) / 8);
+				*sector_mask = 1 << ((sector_in_bank - 32) / 8);
 			}
 			*reg = FCTL_REG_CMDWEPROTB;
 			return ERROR_OK;
